More use of past tense
I had 'is' and such when talking about the blog. While it is still a
blog I don't see it being picked up within the year. Past tense is the
better case with it even if I would like to at some point maybe pick it
back up. Also removed some stuff.