broke a sentence in two and more enters
I had a sentence long enough that in notepad the text went past the edge
of my 1280 wide screen. That is a long sentence. Also since I was bad
and didn't comment about it the previous commit was to make each
sentence its own line so you can see what was done better on github.